refactor: extract dbcache helpers into `node/dbcache.{cpp,h}`
Move the `dbcache`-specific declarations and definitions out of `node/caches.{h,cpp}` into a dedicated `node/dbcache` translation unit.
