+++ src/node/caches.h
@@ -6,19 +6,12 @@
 #define BITCOIN_NODE_CACHES_H
 
 #include <kernel/caches.h>
-#include <util/byte_units.h>
 
 #include <cstddef>
 
 class ArgsManager;
 
-//! min. -dbcache (bytes)
-static constexpr size_t MIN_DB_CACHE{4_MiB};
-//! -dbcache default (bytes)
-static constexpr size_t DEFAULT_DB_CACHE{DEFAULT_KERNEL_CACHE};
-
 namespace node {
-size_t GetDefaultDBCache();
 struct IndexCacheSizes {
     size_t tx_index{0};
     size_t filter_index{0};
@@ -29,11 +22,6 @@ struct CacheSizes {
     kernel::CacheSizes kernel;
 };
 CacheSizes CalculateCacheSizes(const ArgsManager& args, size_t n_indexes = 0);
-constexpr bool ShouldWarnOversizedDbCache(uint64_t dbcache, uint64_t total_ram) noexcept
-{
-    const uint64_t cap{(total_ram < 2_GiB) ? DEFAULT_DB_CACHE : (total_ram / 100) * 75};
-    return dbcache > cap;
-}
 
 void LogOversizedDbCache(const ArgsManager& args) noexcept;
 } // namespace node
